requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==1.26.4
packaging==26.0
pandas==3.0.0
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scikit-surprise==1.1.4
scipy==1.17.0
threadpoolctl==3.6.0
Werkzeug==3.1.5
